Fix debian/bin/kafka
Fix errors in bin/kafka introduced in [1]:
- Use /bin/bash instead of /bin/sh to avoid quoting issues
- Use bash arrays instead of strings for building argument lists
- Expand arg array when calling kafka_run_class

The immediate symptom this fixes is that in cmd_server_start and
cmd_zookeeper_start, arguments to kafka_run_class were collected in a
string variable $EXTRA_ARGS and then passed without expansion to
kafka_run_class, eg.:

    kafka_run_class "${EXTRA_ARGS}" kafka.Kafka "${server_properties}"

Thus, in kafka_run_class, $1 was equal in most cases to the string:

    "-daemon -name kafkaServer -loggc"

This string as a whole, then, was not matched by anything in the case
statement responsible for parsing the arguments separately. Thus, there
was no call to 'shift', and the string was passed to the java invocation
as a part of "$@".

Since -name and -loggc aren't valid options to the java executable, the
command failed.

[1]Iad14021b9c8e7893f44c12b8ec280f5cbb84d4fe

Change-Id: Id096b623144a2fe34d63507f9b76c46183c417a4